Move yaml classes in core module
