+++ kernel/data-pipeline/core/src/main/java/org/apache/shardingsphere/data/pipeline/api/config/process/yaml/swapper/YamlPipelineProcessConfigurationSwapper.java
@@ -15,9 +15,10 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.yaml.process;
+package org.apache.shardingsphere.data.pipeline.api.config.process.yaml.swapper;
 
 import org.apache.shardingsphere.data.pipeline.api.config.process.PipelineProcessConfiguration;
+import org.apache.shardingsphere.data.pipeline.api.config.process.yaml.YamlPipelineProcessConfiguration;
 import org.apache.shardingsphere.infra.util.yaml.swapper.YamlConfigurationSwapper;
 import org.apache.shardingsphere.infra.yaml.config.swapper.algorithm.YamlAlgorithmConfigurationSwapper;
